+++ basalt-core/src/obsidian/config.rs
@@ -1,9 +1,9 @@
-use dirs::config_dir;
+use dirs::{config_dir, home_dir};
 
 use serde::{Deserialize, Deserializer};
 use std::path::Path;
-use std::result;
 use std::{collections::BTreeMap, fs, path::PathBuf};
+use std::{env, result};
 
 use crate::obsidian::{Error, Result, Vault};
 
@@ -19,9 +19,24 @@ impl ObsidianConfig {
     ///
     /// Returns an [`Error`] if the filepath doesn't exist or JSON parsing failed.
     pub fn load() -> Result<Self> {
-        obsidian_config_dir()
-            .map(|path_buf| ObsidianConfig::load_from(&path_buf))
-            .ok_or(Error::PathNotFound("Obsidian config directory".to_string()))?
+        let config_locations = obsidian_global_config_locations();
+        let existing_config_locations = config_locations
+            .iter()
+            .filter(|path| path.is_dir())
+            .collect::<Vec<_>>();
+
+        if let Some(config_dir) = existing_config_locations.first() {
+            ObsidianConfig::load_from(config_dir)
+        } else {
+            Err(Error::PathNotFound(format!(
+                "Obsidian config directory was not found from these locations: {}",
+                config_locations
+                    .iter()
+                    .map(|path| path.to_string_lossy())
+                    .collect::<Vec<_>>()
+                    .join(", ")
+            )))
+        }
     }
 
     /// Attempts to load `obsidian.json` file as an [`ObsidianConfig`] from the given directory
@@ -38,8 +53,18 @@ impl ObsidianConfig {
     /// _ = ObsidianConfig::load_from(Path::new("./dir-with-config-file"));
     /// ```
     pub fn load_from(config_path: &Path) -> Result<Self> {
-        let contents = fs::read_to_string(config_path.join("obsidian.json"))?;
-        Ok(serde_json::from_str(&contents)?)
+        let obsidian_json_path = config_path.join("obsidian.json");
+
+        if obsidian_json_path.try_exists()? {
+            let contents = fs::read_to_string(obsidian_json_path)?;
+            serde_json::from_str(&contents).map_err(Error::Json)
+        } else {
+            // TODO: Maybe a different error should be propagated in this case. E.g. 'unreadable'
+            // file.
+            Err(Error::PathNotFound(
+                obsidian_json_path.to_string_lossy().to_string(),
+            ))
+        }
     }
 
     /// Returns an iterator over the vaults in the configuration.
@@ -183,20 +208,77 @@ impl<'de> Deserialize<'de> for ObsidianConfig {
     }
 }
 
-/// Returns the system path to Obsidian's config folder, if any.
+/// Returns all existing configuration directory paths where Obsidian might store its global
+/// settings.
+///
+/// This function determines possible configuration locations by platform-specific conventions and
+/// installation methods. On all platforms, it first checks if the user has defined the
+/// `OBSIDIAN_CONFIG_DIR` environment variable. If so, that path is used, and any leading tilde (~)
+/// is expanded to the current user's home directory.
+///
+/// On Windows, it then resolves to the default Obsidian directory located under the system's
+/// application data folder, typically `%APPDATA%\Obsidian`. On macOS, the function expects to find
+/// the configuration under `~/Library/Application Support/obsidian`. On Linux, the standard config
+/// directory is assumed to be `$XDG_CONFIG_HOME/obsidian`, or `~/.config/obsidian`.
+///
+/// For Linux users, the function also accounts for sandboxed installations. If Obsidian is
+/// installed via Flatpak, the configuration is likely found in
+/// `~/.var/app/md.obsidian.Obsidian/config/obsidian`. For Snap installations, the relevant path is
+/// typically `~/snap/obsidian/common/.config/obsidian`.
 ///
 /// For reference:
 /// - macOS:  `/Users/username/Library/Application Support/obsidian`
 /// - Windows: `%APPDATA%\Obsidian\`
-/// - Linux:   `$XDG_CONFIG_HOME/obsidian/` or `~/.config/obsidian/`
+/// - Linux:   `$XDG_CONFIG_HOME/obsidian` or `~/.config/obsidian`
+///   flatpak: `$HOME/.var/app/md.obsidian.Obsidian/config/obsidian`
+///   snap: `$HOME/snap/obsidian/common/.config/obsidian`
 ///
 /// More info: [https://help.obsidian.md/Files+and+folders/How+Obsidian+stores+data]
-fn obsidian_config_dir() -> Option<PathBuf> {
+pub fn obsidian_global_config_locations() -> Vec<PathBuf> {
     #[cfg(any(target_os = "macos", target_os = "linux"))]
     const OBSIDIAN_CONFIG_DIR_NAME: &str = "obsidian";
 
     #[cfg(target_os = "windows")]
     const OBSIDIAN_CONFIG_DIR_NAME: &str = "Obsidian";
 
-    config_dir().map(|config_path| config_path.join(OBSIDIAN_CONFIG_DIR_NAME))
+    let override_path =
+        env::var("OBSIDIAN_CONFIG_DIR")
+            .ok()
+            .zip(home_dir())
+            .map(|(path, home_dir)| {
+                PathBuf::from(path.replace("~", home_dir.to_string_lossy().as_ref()))
+            });
+
+    let default_config_path =
+        config_dir().map(|config_path| config_path.join(OBSIDIAN_CONFIG_DIR_NAME));
+
+    #[cfg(any(target_os = "macos", target_os = "windows"))]
+    let sandboxed_paths: [Option<PathBuf>; 0] = [];
+
+    // In cases where user has a sandboxes instance of Obsidian installed under either flatpak or
+    // snap, we must check if the configuration exists under these locations.
+    #[cfg(target_os = "linux")]
+    let sandboxed_paths = {
+        let flatpak_path = home_dir().map(|home_dir| {
+            home_dir
+                .join(".var/app/md.obsidian.Obsidian/config")
+                .join(OBSIDIAN_CONFIG_DIR_NAME)
+        });
+
+        let snap_path = home_dir().map(|home_dir| {
+            home_dir
+                .join("snap/obsidian/common/.config")
+                .join(OBSIDIAN_CONFIG_DIR_NAME)
+        });
+
+        [flatpak_path, snap_path]
+    };
+
+    let base_paths = [override_path, default_config_path];
+
+    base_paths
+        .into_iter()
+        .chain(sandboxed_paths)
+        .flatten()
+        .collect()
 }